@@ -27,16 +27,16 @@ abstract class Builders {
   }
 
   static ListTile buildMenuListTile(
-      {FaIcon icon, String text, Function onTap}) {
+      {FaIcon? icon, required String text, Function? onTap}) {
     return ListTile(
       title: Row(
         children: [
-          icon,
+          icon!,
           SizedBox(width: 10),
           Text(text),
         ],
       ),
-      onTap: onTap,
+      onTap: onTap as void Function()?,
     );
   }
 }
